// src/enums/permission.ts
export enum PermissionEnums {
    MANAGEMENT_FORMS_HUB = 'management_forms_hub',
    MANAGEMENT_CALENDAR = 'management_calendar',
    MANAGEMENT_CONTACTS = 'management_contacts',
    MANAGEMENT_LIBRARY = 'management_library',
    MANAGEMENT_BLUEPRINTS_HUB = 'management_blueprints_hub',
    MANAGEMENT_DEVICE_MANAGEMENT = 'management_device_management',
    MANAGEMENT_USERS_GROUPS = 'management_users_groups',
    MANAGEMENT_SETTINGS = 'management_settings',

    USER_FORMS_HUB = 'user_forms_hub',
    USER_FORM_TEMPLATES = 'user_form_templates',
    USER_FORMS = 'user_forms',
    USER_PUBLIC_FORMS = 'user_public_forms',
    USER_DELIVERY_ACTIVITY = 'user_delivery_activity',
    USER_CALENDAR = 'user_calendar',
    USER_CONTACTS = 'user_contacts',
    USER_LIBRARY = 'user_library',
    USER_BLUEPRINTS_HUB = 'user_blueprints_hub',
    USER_PROJECTS = 'user_projects',

    ADVANCED_FORMS_HUB_EXPORT_PDF = 'advanced_forms_hub_export_pdf',
    ADVANCED_FORMS_HUB_ADD_TAGS = 'advanced_forms_hub_add_tags',
    ADVANCED_FORMS_HUB_RENAME = 'advanced_forms_hub_rename',
    ADVANCED_FORMS_HUB_DELETE = 'advanced_forms_hub_delete',
    ADVANCED_FORMS_HUB_CREATE_COPIES = 'advanced_forms_hub_create_copies',
    ADVANCED_FORMS_HUB_DELETE_PUBLIC = 'advanced_forms_hub_delete_public',

    ADVANCED_BLUEPRINTS_HUB_EXPORT_PDF = 'advanced_blueprints_hub_export_pdf',
    ADVANCED_BLUEPRINTS_HUB_ADD_TAGS_PROJECT = 'advanced_blueprints_hub_add_tags_project',
    ADVANCED_BLUEPRINTS_HUB_RENAME_PROJECT = 'advanced_blueprints_hub_rename_project',
    ADVANCED_BLUEPRINTS_HUB_DELETE_PROJECT = 'advanced_blueprints_hub_delete_project',
    ADVANCED_BLUEPRINTS_HUB_EMAIL_PLANS = 'advanced_blueprints_hub_email_plans',
    ADVANCED_BLUEPRINTS_HUB_DELETE_ANNOTATIONS = 'advanced_blueprints_hub_delete_annotations',
    ADVANCED_BLUEPRINTS_HUB_CHANGE_STATUS = 'advanced_blueprints_hub_change_status',
}